Fix compile issue: text.cpp's compiled .obj file was being overwritten by ada/tools/text.cpp.  Rename ada/tools/text.cpp to ada_text.cpp
