+++ src/ds/util/ada/tools/ada_text.h
